add mergehandle interface
