+++ hudi-client/hudi-flink-client/src/main/java/org/apache/hudi/table/action/commit/BaseFlinkCommitActionExecutor.java
@@ -32,7 +32,7 @@ import org.apache.hudi.exception.HoodieUpsertException;
 import org.apache.hudi.execution.FlinkLazyInsertIterable;
 import org.apache.hudi.io.ExplicitWriteHandleFactory;
 import org.apache.hudi.io.HoodieCreateHandle;
-import org.apache.hudi.io.HoodieDefaultMergeHandle;
+import org.apache.hudi.io.HoodieWriteMergeHandle;
 import org.apache.hudi.io.HoodieWriteHandle;
 import org.apache.hudi.table.HoodieTable;
 import org.apache.hudi.table.action.HoodieWriteMetadata;
@@ -166,7 +166,7 @@ public abstract class BaseFlinkCommitActionExecutor<T> extends
         // the second batch batch2 tries to reuse the same bucket
         // and append instead of UPDATE.
         return handleInsert(fileIdHint, recordItr);
-      } else if (this.writeHandle instanceof HoodieDefaultMergeHandle) {
+      } else if (this.writeHandle instanceof HoodieWriteMergeHandle) {
         return handleUpdate(partitionPath, fileIdHint, recordItr);
       } else {
         switch (bucketType) {
@@ -190,17 +190,17 @@ public abstract class BaseFlinkCommitActionExecutor<T> extends
                                                   Iterator<HoodieRecord<T>> recordItr)
       throws IOException {
     // This is needed since sometimes some buckets are never picked in getPartition() and end up with 0 records
-    HoodieDefaultMergeHandle<?, ?, ?, ?> upsertHandle = (HoodieDefaultMergeHandle<?, ?, ?, ?>) this.writeHandle;
+    HoodieWriteMergeHandle<?, ?, ?, ?> upsertHandle = (HoodieWriteMergeHandle<?, ?, ?, ?>) this.writeHandle;
     if (upsertHandle.isEmptyNewRecords() && !recordItr.hasNext()) {
       LOG.info("Empty partition with fileId => {}.", fileId);
       return Collections.singletonList((List<WriteStatus>) Collections.EMPTY_LIST).iterator();
     }
     // these are updates
-    HoodieDefaultMergeHandle<?, ?, ?, ?> mergeHandle = (HoodieDefaultMergeHandle<?, ?, ?, ?>) this.writeHandle;
+    HoodieWriteMergeHandle<?, ?, ?, ?> mergeHandle = (HoodieWriteMergeHandle<?, ?, ?, ?>) this.writeHandle;
     return handleUpdateInternal(mergeHandle, fileId);
   }
 
-  protected Iterator<List<WriteStatus>> handleUpdateInternal(HoodieDefaultMergeHandle<?, ?, ?, ?> mergeHandle, String fileId)
+  protected Iterator<List<WriteStatus>> handleUpdateInternal(HoodieWriteMergeHandle<?, ?, ?, ?> mergeHandle, String fileId)
       throws IOException {
     if (mergeHandle.getOldFilePath() == null) {
       throw new HoodieUpsertException(
